requirements.txt
streamlit>=1.20
opencv-python
pillow
numpy
matplotlib
fer
